Avoid testing if a tab is actually closed
If Chrome returned 200 OK and the tab is not actually closed it is not
a problem of this module, as long as the invocation "/json/close/<id>"
is performed correctly. On the other hand Chrome responds !200 in case
of wrong URLs. This is a tradeoff; a timeout in this context feels too
fragile to me.