Make the 'close()' call properly asynchronous
Now it accepts a callback as an argument and as usual it returns a Promise
object if the callback is omitted. All the test cases have been updated in
order to use this feature. Now there should be no more race conditions about
WebSocket termination and subsequent operations.

This should fix #38.