Use the local version of the protocol by default
Now the users have to explicitly state their intent of accessing the Internet
(this concerns Chrome and not Edge) to fetch the exact version of the protocol
for the currently used browser.

The default value of `fallback` has been inverted; in doing so the `fallback`
option is now named `remote`. This is a possibly breaking change.