Alter the protocol fetching behavior
Since Chrome 60 (now old stable) it is possible to fetch the protocol from the
remote instance directly, so there is no more need of fetching it from the
repository. It makes sense thus to make it the default behavior, in doing so the
`remote` option became the `local` option.

The FAQ changed accordingly.

Moreover the `Protocol` *static* method now returns just the descriptor, i.e.,
not an additional `remote` flag.

These are breaking changes.

Close #264.