TypeScript mentions and example tweak.
I tried your late assignment and it broke the implicit type sensing in vscode. So I modified it to still be implicitly sensed. Unfortunately, putting the catch on makes it possible for the assignment to be void so it did end up needing to be cast for Typescript. Turns out there is more missing from the DefinitelyTyped definition than I thought. So your event handlers are only defined with the standard emitter pattern not your eventname as a method with a callback.